Move mark done tx helpers to separate file
